+++ EquipmentAndQuickSlots/EquipmentAndQuickSlots.cs
@@ -22,7 +22,7 @@ namespace EquipmentAndQuickSlots {
     [BepInIncompatibility("com.bruce.valheim.comfyquickslots")]
     public class EquipmentAndQuickSlots : BaseUnityPlugin {
         public const string PluginId = "randyknapp.mods.equipmentandquickslots";
-        public const string Version = "3.1.2";
+        public const string Version = "3.1.3";
 
         public static Sprite PaperdollMale;
         public static Sprite PaperdollFemale;